backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
claude-code-sdk
pdfminer.six==20231228
pytesseract==0.3.10
pillow==10.4.0
python-multipart==0.0.9
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.0
e2b==0.17.1
playwright==1.40.0
jinja2==3.1.2